Update tests, removed newly created test class and added tests into old one.
